% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{.add_assignment_constraints}
\alias{.add_assignment_constraints}
\title{Add Assignment Constraints to MILP}
\usage{
.add_assignment_constraints(model, ps_by_i, n)
}
\arguments{
\item{model}{An ompr MILP model.}

\item{ps_by_i}{List mapping tract index to sparse-pair indices.}

\item{n}{Integer. Number of tracts.}
}
\value{
Updated MILP model with assignment constraints.
}
\description{
Each tract must be assigned to exactly one center.
}
\keyword{internal}
