% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{.is_symmetric_matrix}
\alias{.is_symmetric_matrix}
\title{Check if a Matrix is Symmetric}
\usage{
.is_symmetric_matrix(mat, tol = 1e-08)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{tol}{Tolerance for floating point comparison.}
}
\value{
Logical scalar.
}
\description{
Check if a Matrix is Symmetric
}
\keyword{internal}
