% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{.matrix_to_sparse}
\alias{.matrix_to_sparse}
\title{Convert a Symmetric Distance Matrix to Sparse Format}
\usage{
.matrix_to_sparse(mat)
}
\arguments{
\item{mat}{Symmetric numeric matrix with IDs as row/col names.}
}
\value{
A tibble with columns \code{origin_id}, \code{destination_id}, \code{distance}.
}
\description{
Convert a Symmetric Distance Matrix to Sparse Format
}
\keyword{internal}
