% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\name{.seriate_quietly}
\alias{.seriate_quietly}
\title{Call seriation::seriate() Suppressing Only TSP Control Warnings}
\usage{
.seriate_quietly(d, method, control)
}
\arguments{
\item{d}{A \code{dist} object.}

\item{method}{Seriation method.}

\item{control}{Control list.}
}
\value{
A seriation order object.
}
\description{
Call seriation::seriate() Suppressing Only TSP Control Warnings
}
\keyword{internal}
