% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{surveyzones_plot_zones}
\alias{surveyzones_plot_zones}
\title{Plot Zone Assignments}
\usage{
surveyzones_plot_zones(
  plan,
  tracts_sf,
  show_centers = TRUE,
  show_sequence = FALSE
)
}
\arguments{
\item{plan}{A \code{surveyzones_plan} object.}

\item{tracts_sf}{An sf object with a \code{tract_id} column and polygon
(or point) geometries.}

\item{show_centers}{Logical.  Highlight zone centers? Default \code{TRUE}.}

\item{show_sequence}{Logical.  Draw visit-order arrows?
Requires that \code{\link[=surveyzones_sequence]{surveyzones_sequence()}} has been called on \code{plan}.
Default \code{FALSE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise zone assignments on a map.  Requires tract geometries
to be provided as an sf object.
}
