% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{validate_access_points}
\alias{validate_access_points}
\title{Validate Access Points}
\usage{
validate_access_points(access_points, call = rlang::caller_env())
}
\arguments{
\item{access_points}{Object to validate.}

\item{call}{Caller environment for error reporting.}
}
\value{
\code{access_points}, invisibly.
}
\description{
Validate Access Points
}
\keyword{internal}
