% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{validate_tracts}
\alias{validate_tracts}
\title{Validate Tract Table}
\usage{
validate_tracts(tracts, call = rlang::caller_env())
}
\arguments{
\item{tracts}{Object to validate.}

\item{call}{Caller environment for error reporting.}
}
\value{
\code{tracts}, invisibly.
}
\description{
Validate Tract Table
}
\keyword{internal}
